final_project/myapp_test/requirements.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
apipkg==1.5
assertpy==1.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
docker==5.0.0
execnet==1.8.1
Faker==8.5.0
greenlet==1.1.0
idna==2.10
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.4
pytest-check==1.0.1
pytest-forked==1.3.0
pytest-xdist==2.2.1
python-dateutil==2.8.1
requests==2.25.1
selenium==3.141.0
six==1.16.0
SQLAlchemy==1.4.17
text-unidecode==1.3
toml==0.10.2
url-normalize==1.4.3
urllib3==1.26.5
websocket-client==1.0.1
